api/requirements.txt
fastapi==0.99.1
fastapi_jwt_auth==0.5.0
websockets==11.0.3
uvicorn==0.23.2
python-multipart==0.0.9
requests
pydantic==1.10.16
pillow
pyotp
pyjwt
